+++ app/src/main/java/com/brandonjamesyoung/levelup/data/PurchasedItem.kt
@@ -6,9 +6,10 @@ import androidx.room.PrimaryKey
 import java.time.Instant
 
 @Entity
-data class Item(
+data class PurchasedItem(
     @PrimaryKey(autoGenerate = true) var id: Int = 0,
     @ColumnInfo val name: String? = null,
     @ColumnInfo val cost: Int,
-    @ColumnInfo val dateCreated: Instant? = Instant.now()
+    @ColumnInfo val dateCreated: Instant?,
+    @ColumnInfo val datePurchased: Instant = Instant.now()
 )